+++ src/components/common/ContainedButton.tsx
@@ -10,6 +10,7 @@ interface ContainedButtonProps {
   disabled?: boolean;
   type?: "button" | "submit" | "reset";
   icon?: React.ReactNode;
+  className?: string;
 }
 
 export const ContainedButton = ({
@@ -20,13 +21,16 @@ export const ContainedButton = ({
   disabled = false,
   type = "button",
   icon,
+  className,
 }: ContainedButtonProps) => {
   return (
     <button
       onClick={handleClick}
       type={type}
       disabled={disabled || loading}
-      className="transition text-nowrap w-full h-full flex items-center justify-center disabled:opacity-75 bg-containedButtonBg dark:bg-containedButtonBgDark dark:hover:bg-containedButtonBgHoverDark hover:bg-containedButtonBgHover text-white p-2 rounded-md focus:outline-none"
+      className={`transition text-nowrap w-full h-full flex items-center justify-center disabled:opacity-75 bg-containedButtonBg 
+      dark:bg-containedButtonBgDark dark:hover:bg-containedButtonBgHoverDark hover:bg-containedButtonBgHover 
+      text-white p-2 rounded-md focus:outline-none ${className}`}
     >
       {icon && <div className="mr-2"> {icon}</div>}
       {loading ? <SpinnerIcon /> : children || text}