+++ src/layout/sideMenu/Logo.tsx
@@ -9,8 +9,8 @@ export const Logo = () => {
 
   return (
     <div
-      className={`-ml-4 h-20  
-        text-mainColor dark:text-mainColorDark dark:fill-mainColorDark text-center flex justify-center items-center text-[1.4rem] xl:text-[1.2rem] 2xl:text-[1.3rem] font-bold ${nunito.className}`}
+      className={`-ml-1 h-20  
+        text-mainColor dark:text-mainColorDark dark:fill-mainColorDark text-center flex justify-center items-center text-[1.4rem] xl:text-[1.2rem] 3xl:text-[1.3rem] font-bold ${nunito.className}`}
     >
       <div className="-mt-1">
         <LogoIcon />